+++ src/game/player/lib/Inspect.tsx
@@ -1,6 +1,7 @@
 import React, {useState} from "react";
 import {DivElementProp} from "@core/elements/transition/type";
 import {useGame} from "@player/provider/game-state";
+import {m} from "framer-motion";
 
 type InspectStyle = {
     border?: "solid" | "dashed" | "dotted";
@@ -25,11 +26,13 @@ function InspectBase<T extends keyof React.JSX.IntrinsicElements | React.Compone
         tag,
         borderWidth = 1,
         as: Component = "div",
+        Ref,
         ...props
     }: Readonly<DivElementProp & {
         children?: React.ReactNode;
         tag?: string;
         as?: T;
+        Ref?: React.RefObject<HTMLDivElement>;
     } & InspectStyle & ElementProps<T>>) {
     const {game} = useGame();
     const [isHovered, setIsHovered] = useState(false);
@@ -50,14 +53,61 @@ function InspectBase<T extends keyof React.JSX.IntrinsicElements | React.Compone
     };
 
     return (
-        <Component {...commonProps}>
+        <Component {...commonProps} ref={Ref}>
             {tag && isHovered && (
-                <span className="absolute top-0 left-0 bg-white text-black border-2 border-black">
+                <span className="absolute top-0 left-0 bg-white text-black border-2 border-black text-sm">
                     {tag}
                 </span>
             )}
             {children}
         </Component>
+    );
+}
+
+/**
+ * For self-closing tags
+ * @constructor
+ */
+function InspectCloseBase<T extends keyof React.JSX.IntrinsicElements | React.ComponentType<any>>(
+    {
+        border = "solid",
+        color = "red",
+        tag,
+        borderWidth = 1,
+        as: Component = "img",
+        ...props
+    }: Readonly<DivElementProp & {
+        tag?: string;
+        as?: T;
+    } & InspectStyle & ElementProps<T>>
+) {
+    const {game} = useGame();
+    const [isHovered, setIsHovered] = useState(false);
+
+    if (!game.config.app.inspector) {
+        return <Component {...props}/>;
+    }
+
+    const commonProps = {
+        ...props,
+        onMouseEnter: () => setIsHovered(true),
+        onMouseLeave: () => setIsHovered(false),
+        style: {
+            ...(props.style || {}),
+            border: `${border} ${borderWidth}px ${color}`,
+            zIndex: isHovered ? 1000 : "auto",
+        },
+    };
+
+    return (
+        <div>
+            <Component {...commonProps} />
+            {tag && isHovered && (
+                <span className="absolute top-0 left-0 bg-white text-black border-2 border-black">
+                    {tag}
+                </span>
+            )}
+        </div>
     );
 }
 
@@ -69,6 +119,10 @@ function InspectSpan(props: Readonly<DivElementProp & { children?: React.ReactNo
     return <InspectBase {...props} as="span"/>;
 }
 
+function InspectImg(props: Readonly<DivElementProp & { tag?: string; } & InspectStyle>) {
+    return <InspectCloseBase {...props} as="img"/>;
+}
+
 function InspectButton(props: Readonly<DivElementProp & {
     children?: React.ReactNode;
     tag?: string;
@@ -76,10 +130,20 @@ function InspectButton(props: Readonly<DivElementProp & {
     return <InspectBase {...props} as="button"/>;
 }
 
+function InspectFramerMotionDiv(props: Readonly<DivElementProp & {
+    children?: React.ReactNode;
+    tag?: string;
+    Ref?: React.RefObject<HTMLDivElement>;
+} & InspectStyle>) {
+    return <InspectBase {...props} as={m.div} Ref={props.Ref}/>;
+}
+
 const Inspect = {
     Div: InspectDiv,
     Span: InspectSpan,
     Button: InspectButton,
+    Img: InspectImg,
+    mDiv: InspectFramerMotionDiv,
 };
 
 export default Inspect;